TrainManagerServer/requirements.txt
fastapi[standard]=0.115.8
pyjwt=2.10.1
passlib=1.7.4
bcrypt=4.2.1
SQLAlchemy=2.0.38
pydantic=2.10.6
argon2_cffi=23.1.0
httpx=0.28.1
sqlmodel=0.0.22 